Handle external protocols with the network service.
https://chromium-review.googlesource.com/c/chromium/src/+/956091
